+++ SumNode.cs
@@ -2,7 +2,7 @@
 {
     public class SumNode : ProcessingNode
     {
-        public SumNode(Func<double, double> activationFunc = null, float? depth = null) : base(activationFunc, depth) { }
+        public SumNode(Func<double, double> activationFunc = null, double? depth = null) : base(activationFunc, depth) { }
 
         protected override double AggregateInput(List<Synapse> inputs)
         {
